examples/bl-hello-app-template/requirements.txt
fake-bpy-module-3.3